+++ angular/src/app/supply/supply-ulist/supply-ulist.component.ts
@@ -2,7 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { Suplly } from 'src/app/model/supply';
 import { Router } from '@angular/router';
 import { SupplyService } from '../supply.service';
-
+import Swal from 'sweetalert2';
 @Component({
   selector: 'app-supply-ulist',
   templateUrl: './supply-ulist.component.html',
@@ -25,25 +25,41 @@ export class SupplyUListComponent implements OnInit {
 }
 delete(id:number)
 {
-  this.supplyService.delete(id).subscribe(
-    (res)=>
-    {
-      this.dataSource=this.dataSource.filter(item=>item.carId!=id);
-      alert("ok");
-    },
-    (err)=>
-    {
-      alert("err");
-    }
-  );
+  Swal.fire({
+    title: 'את/ה בטוח/ה ',
+    text: "לא תוכל/י לבטל!",
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonColor: '#3085d6',
+    cancelButtonColor: '#d33',
+    confirmButtonText: 'כן,מחק!'
+  }).then((result) => {
+    if (result.value) {
+      this.supplyService.delete(id).subscribe(
+      (res)=>
+      {
+        Swal.fire(
+          'נמחק',
+          'בקשתך נמחקה.',
+          'success'
+        );
+        this.dataSource=this.dataSource.filter(item=>item.supplyId!=id);
+      },
+      (err)=>
+      {
+        alert("err");
+      }
+    );
+      
+  }
+    
+  })
+    
 }
 supplyDetails(supply:Suplly)
 {
  return this.router.navigate(['supply-details']);
  
 }
-
-
-
 
 }
